these changes are made to better sync with the approach to be used in Waldo_FCS.
changes include calling the PosVel request from a 10Hz timer. This
emulates the call to the 10Hz timer that will be used for Waldo_FCS. The
Camera trigger request is also called from the same timer but at a
slower interval. There are now three threads: requestPosVel,
requestTrigger, and catch the image file name as recorded by the Canon
camera.