removed the posvel and trigger threads and replaced with inline code. Added retrieval of the SD card file at completion
replaced the threaded methods and timer with a single while() where
everything runs as fast as possible. The posvel request is made and we
wait for the response. A sleet is used to represent the real-time
action. Then we use a timer to invoke a trigger request (e.g., 10 sec
period). I see that we can get reliable mbed transfer with this method
of several 1000 posvel requests. Can slow this down with the Sleep().
Also added the code from Sam that downloads the SD card file.